// src/generated/models/Dga_technologiesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Dga_technologiesdga_technologytype = {
  1: 'Product',
  2: 'Company'
} as const;
export type Dga_technologiesdga_technologytype = keyof typeof Dga_technologiesdga_technologytype;
export const Dga_technologiesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Dga_technologiesstatecode = keyof typeof Dga_technologiesstatecode;
export const Dga_technologiesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Dga_technologiesstatuscode = keyof typeof Dga_technologiesstatuscode;

export interface Dga_technologiesBase {
  dga_technology_name?: string;
  dga_technologyid: string;
  dga_technologytype?: Dga_technologiesdga_technologytype;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  "OwningBusinessUnit@odata.bind"?: string;
  statecode: Dga_technologiesstatecode;
  statuscode?: Dga_technologiesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Dga_technologies extends Dga_technologiesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  dga_technologytypename?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}